requirements.txt
Flask==3.0.3
psycopg==3.2.13
psycopg-binary==3.2.13
Werkzeug==3.0.3
python-dotenv==1.0.1
